update modal and text components
