+++ ui/app/pages/workspaces/workspaces.main.tsx
@@ -2,13 +2,8 @@ import React, { Suspense, useState } from 'react';
 
 import { PlusCircleIcon } from '@heroicons/react/20/solid';
 import { MagnifyingGlassIcon } from '@heroicons/react/24/solid';
-import {
-  useMutation,
-  useQuery,
-  useQueryClient,
-  UseQueryOptions,
-} from 'react-query';
-import { Await, useLoaderData, useParams } from 'react-router-dom';
+import { useMutation, useQuery, useQueryClient, UseQueryOptions } from 'react-query';
+import { Await, useLoaderData } from 'react-router-dom';
 
 import {
   createWorkspace,
@@ -67,9 +62,13 @@ export const useUpdateWorkspace = (instanceKey: string | undefined) => {
       description: string;
       tags: string[];
     }) => {
-      await configureAxios(instanceKey!);
+      if (!instanceKey || !workspaceKey) {
+        throw new Error('instanceKey is undefined');
+      }
+
+      await configureAxios(instanceKey);
       await updateWorkspace({
-        workspaceKey: workspaceKey!,
+        workspaceKey,
         body: [
           {
             op: 'replace',
@@ -134,42 +133,32 @@ export const useRemoveWorkspace = (instanceKey: string | undefined) => {
 export const useWorkspaces = (options?: UseQueryOptions<Workspace[]>) => {
   const { instanceKey } = useFlagbaseParams();
   const { data: instances } = useInstances({
-    select: (instances) =>
-      instances.filter((instance) => instance.key === instanceKey),
+    select: (instances) => instances.filter((instance) => instance.key === instanceKey),
   });
-  const query = useQuery<Workspace[]>(
-    ['workspaces', instanceKey?.toLocaleLowerCase()],
-    {
-      queryFn: async () => {
-        if (instanceKey) {
-          await configureAxios(instanceKey);
+  const query = useQuery<Workspace[]>(['workspaces', instanceKey?.toLocaleLowerCase()], {
+    queryFn: async () => {
+      if (instanceKey) {
+        await configureAxios(instanceKey);
 
-          return fetchWorkspaces();
-        }
-        throw new Error('instanceKey is undefined');
-      },
-      enabled: instances && instances.length > 0,
-      ...options,
+        return fetchWorkspaces();
+      }
+      throw new Error('instanceKey is undefined');
     },
-  );
+    enabled: instances && instances.length > 0,
+    ...options,
+  });
 
   return query;
 };
 
 const MainWorkspaces = () => {
-  const { instanceKey } = useParams() as { instanceKey: string };
   const { instance, workspaces: prefetchedWorkspaces } = useLoaderData() as {
     workspaces: Workspace[];
     instance: Instance;
   };
   const [createWorkspace, showCreateWorkspace] = useState(false);
   const [filter, setFilter] = useState('');
-  const {
-    data: workspaces,
-    isRefetching,
-    isFetching,
-    isLoading,
-  } = useWorkspaces(instanceKey);
+  const { data: workspaces, isRefetching, isFetching, isLoading } = useWorkspaces();
 
   return (
     <Suspense fallback={<Loader />}>
@@ -196,13 +185,7 @@ const MainWorkspaces = () => {
             <Table
               loading={isFetching || isRefetching || isLoading}
               dataSource={
-                workspaces
-                  ? convertWorkspaces(
-                      workspaces,
-                      instance,
-                      filter.toLowerCase(),
-                    )
-                  : []
+                workspaces ? convertWorkspaces(workspaces, instance, filter.toLowerCase()) : []
               }
               columns={workspaceColumns}
               emptyState={
