+++ ui/components/layout/layout.tsx
@@ -1,13 +1,4 @@
-import React, { Fragment } from 'react';
-
-import { Dialog, Transition } from '@headlessui/react';
-
-export type LayoutProps = {
-  open: boolean;
-  onClose: () => void;
-  children: React.ReactChild;
-  className?: string;
-};
+import React from 'react';
 
 const Layout: React.FC = ({ children }) => {
   return (
@@ -15,53 +6,4 @@ const Layout: React.FC = ({ children }) => {
   );
 };
 
-const ModalLayout: React.FC<LayoutProps> = ({
-  open,
-  onClose,
-  children,
-  className,
-}) => {
-  return (
-    <Transition.Root show={open} as={Fragment}>
-      <Dialog as="div" className="relative z-10" onClose={onClose}>
-        <Transition.Child
-          as={Fragment}
-          enter="ease-out duration-300"
-          enterFrom="opacity-0"
-          enterTo="opacity-100"
-          leave="ease-in duration-200"
-          leaveFrom="opacity-100"
-          leaveTo="opacity-0"
-        >
-          <div className="fixed inset-0 bg-gray-500 bg-opacity-75 transition-opacity" />
-        </Transition.Child>
-
-        <div className="fixed inset-0 z-10 overflow-y-auto">
-          <div className="flex min-h-full items-end justify-center p-4 text-center sm:items-center sm:p-0">
-            <Transition.Child
-              as={Fragment}
-              enter="ease-out duration-300"
-              enterFrom="opacity-0 translate-y-4 sm:translate-y-0 sm:scale-95"
-              enterTo="opacity-100 translate-y-0 sm:scale-100"
-              leave="ease-in duration-200"
-              leaveFrom="opacity-100 translate-y-0 sm:scale-100"
-              leaveTo="opacity-0 translate-y-4 sm:translate-y-0 sm:scale-95"
-            >
-              <Dialog.Panel
-                className={
-                  className
-                    ? className
-                    : 'relative transform overflow-hidden rounded-lg bg-white px-4 pt-5 pb-4 text-left shadow-xl transition-all sm:my-8 sm:w-full sm:max-w-sm sm:p-6'
-                }
-              >
-                {children}
-              </Dialog.Panel>
-            </Transition.Child>
-          </div>
-        </div>
-      </Dialog>
-    </Transition.Root>
-  );
-};
-
-export { Layout, ModalLayout };
+export { Layout };
