+++ ui/components/text/text.tsx
@@ -2,10 +2,9 @@ import React from 'react';
 
 export type TextProps = {
   children: React.ReactChild;
+  className?: string;
 };
 
-const Text: React.FC<TextProps> = (props) => {
+export const Text: React.FC<TextProps> = (props) => {
   return <span {...props} />;
 };
-
-export default Text;
