+++ ui/package.json
@@ -126,6 +126,7 @@
     "react-draggable": "^4.4.5",
     "react-query": "^3.39.3",
     "react-router-dom": "^6.8.2",
+    "react-table": "^7.8.0",
     "rfc6902": "^5.0.1",
     "tsconfig-paths": "^4.1.0",
     "yup": "^0.32.11"
